backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.32.0